requirements.txt
streamlit
openai
python-dotenv
pypdf
langchain
langchain-openai
langchain-community
langchain-classic
chromadb
tiktoken